Extract virt and integration tests to libwebauthn-tests
