+++ libwebauthn-tests/src/virt/data/fido-cert.der
