+++ libwebauthn-tests/src/virt/data/fido-key.trussed
