+++ libwebauthn-tests/src/virt/mod.rs
@@ -1,31 +1,26 @@
 mod device;
 mod pipe;
 
-use super::hid::framing::HidCommand;
-use super::hid::framing::HidMessage;
-use crate::webauthn::Error;
-use num_enum::TryFromPrimitive;
 use std::sync::mpsc::{Receiver, Sender};
 use std::thread;
 use std::thread::JoinHandle;
 
+use libwebauthn::proto::CtapError;
+use libwebauthn::transport::hid::framing::{HidCommand, HidMessage};
+use libwebauthn::transport::hid::{virtual_device, HidDevice, HidPipeBackend};
+use num_enum::TryFromPrimitive;
+
+/// `HidPipeBackend` implementation backed by an in-process trussed-staging
+/// fido-authenticator. Each instance owns a worker thread that owns the
+/// trussed client; messages are exchanged over `mpsc` channels.
 #[derive(Debug)]
-pub(crate) struct VirtHidDevice {
+pub struct TrussedVirtBackend {
     req_tx: Sender<HidMessage>,
     resp_rx: Receiver<HidMessage>,
-    _device_thread: JoinHandle<()>,
+    _worker: JoinHandle<()>,
 }
 
-impl Drop for VirtHidDevice {
-    fn drop(&mut self) {
-        // Telling the thread to quit
-        let _ = self
-            .req_tx
-            .send(HidMessage::new(0, HidCommand::Cancel, &[]));
-    }
-}
-
-impl VirtHidDevice {
+impl TrussedVirtBackend {
     pub fn new() -> Self {
         let (req_tx, req_rx) = std::sync::mpsc::channel::<HidMessage>();
         let (resp_tx, resp_rx) = std::sync::mpsc::channel::<HidMessage>();
@@ -34,18 +29,14 @@ impl VirtHidDevice {
         // device::run_ctaphid() multiple times, as that would mean a new
         // device initialization for each call (and rendering every shared
         // secret invalid).
-        // So we let it run only once, but in a seperate thread, and simply
+        // So we let it run only once, but in a separate thread, and simply
         // pass messages back and forth.
-        // This way, the device works across multiple request calls,
-        // as long as the device-object lives. Everything is wiped on dropping
-        // the device object.
-        let thread_handle = thread::spawn(move || {
+        let worker = thread::spawn(move || {
             device::run_ctaphid(move |device| {
                 while let Ok(msg) = req_rx.recv() {
                     let resp = match msg.cmd {
                         HidCommand::Ping => device.ping(&msg.payload).map(|_| Vec::new()),
                         HidCommand::Msg => device.ctap1(&msg.payload),
-                        // HidCommand::Lock => device.lock(duration),
                         HidCommand::Init => {
                             let mut payload = msg.payload.clone();
                             // Fake channel ID
@@ -59,17 +50,14 @@ impl VirtHidDevice {
                             Ok(payload)
                         }
                         HidCommand::Wink => device.wink().map(|_| Vec::new()),
-                        HidCommand::Cbor => {
-                            device
-                                .ctap2(msg.payload[0], &msg.payload[1..])
-                                .map(|payload| {
-                                    // For CBOR, we have to put the status code in front.
-                                    // If we get here, it was successful, so we add:
-                                    let mut status_with_payload = vec![0]; // Status code: Ok
-                                    status_with_payload.extend_from_slice(&payload);
-                                    status_with_payload
-                                })
-                        }
+                        HidCommand::Cbor => device
+                            .ctap2(msg.payload[0], &msg.payload[1..])
+                            .map(|payload| {
+                                // For CBOR, status code goes in front; success = 0.
+                                let mut status_with_payload = vec![0];
+                                status_with_payload.extend_from_slice(&payload);
+                                status_with_payload
+                            }),
                         HidCommand::Cancel => break,
                         HidCommand::Lock
                         | HidCommand::Sync
@@ -80,25 +68,20 @@ impl VirtHidDevice {
                         Ok(payload) => {
                             let mut response = msg.clone();
                             response.payload = payload;
-                            match resp_tx.send(response) {
-                                Ok(_) => continue,
-                                Err(_) => break,
+                            if resp_tx.send(response).is_err() {
+                                break;
                             }
                         }
                         Err(ctaphid::error::Error::CommandError(
                             ctaphid::error::CommandError::CborError(value),
-                        )) => match crate::proto::CtapError::try_from_primitive(value) {
+                        )) => match CtapError::try_from_primitive(value) {
                             Ok(_) => {
-                                // If we have a known Error status code, we return `Ok(())`
-                                // (because the transmission was successful, but the operation was not)
-                                // and let the code above us handle the error accordingly
-                                // on a subsequent recv-call
+                                // Known CTAP error code: forward as a successful
+                                // transmission with the status byte as payload.
                                 let mut response = msg.clone();
-                                let status_with_payload = vec![value]; // Status code: Err
-                                response.payload = status_with_payload; // No additional payload
-                                match resp_tx.send(response) {
-                                    Ok(_) => continue,
-                                    Err(_) => break,
+                                response.payload = vec![value];
+                                if resp_tx.send(response).is_err() {
+                                    break;
                                 }
                             }
                             Err(_) => panic!("Failed to parse CtapError from {value}"),
@@ -108,20 +91,41 @@ impl VirtHidDevice {
                 }
             })
         });
+
         Self {
-            _device_thread: thread_handle,
             req_tx,
             resp_rx,
+            _worker: worker,
         }
     }
+}
 
-    pub fn virt_send(&mut self, msg: &HidMessage) -> Result<(), Error> {
+impl Default for TrussedVirtBackend {
+    fn default() -> Self {
+        Self::new()
+    }
+}
+
+impl Drop for TrussedVirtBackend {
+    fn drop(&mut self) {
+        // Tell the worker thread to exit cleanly.
+        let _ = self
+            .req_tx
+            .send(HidMessage::new(0, HidCommand::Cancel, &[]));
+    }
+}
+
+impl HidPipeBackend for TrussedVirtBackend {
+    fn send(&mut self, msg: &HidMessage) {
         let _ = self.req_tx.send(msg.to_owned());
-        Ok(())
     }
 
-    pub fn virt_recv(&mut self) -> Result<HidMessage, Error> {
-        let response = self.resp_rx.recv().unwrap();
-        Ok(response)
+    fn recv(&mut self) -> HidMessage {
+        self.resp_rx.recv().expect("virt worker disconnected")
     }
+}
+
+/// Convenience constructor matching the previous `get_virtual_device()` helper.
+pub fn get_virtual_device() -> HidDevice {
+    virtual_device(TrussedVirtBackend::new())
 }
