--- libwebauthn/src/tests/mod.rs
@@ -1,4 +1,0 @@
-pub mod basic_ctap1;
-pub mod basic_ctap2;
-pub mod pin_protocols;
-pub mod prf;
