+++ libwebauthn/src/transport/hid/mod.rs
@@ -5,8 +5,8 @@ pub mod device;
 pub mod framing;
 pub mod init;
 
-#[cfg(test)]
-pub use device::get_virtual_device;
+#[cfg(feature = "virt")]
+pub use device::{virtual_device, HidPipeBackend};
 pub use device::{list_devices, HidDevice};
 
 use super::Transport;
